文件编码

+++ basicKnowledge/src/com/xdc/basic/tools/file2string/String2File.java
@@ -1,13 +1,17 @@
 package com.xdc.basic.tools.file2string;
 
+import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.FileReader;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.io.Reader;
+import java.util.Collection;
 import java.util.List;
 
 import org.apache.commons.codec.binary.Base64;
+import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang3.StringUtils;
 
@@ -20,8 +24,18 @@ public class String2File
 {
     public static void main(String[] args) throws IOException
     {
-        String fromFileName = "data.txt";
+        String curPath = GetPath.getRelativePath();
+        Collection<File> txtFiles = FileUtils.listFiles(new File(curPath), new String[] { "txt" }, false);
+        for (File txtFile : txtFiles)
+        {
+            String fromFileName = txtFile.getName();
 
+            string2File(fromFileName);
+        }
+    }
+
+    private static void string2File(String fromFileName) throws FileNotFoundException, IOException
+    {
         String curPath = GetPath.getRelativePath();
 
         // 读取文件
@@ -47,6 +61,16 @@ public class String2File
         // 解析内容
         String toFileContentBase64String = StringUtils.substringAfter(toFileBase64String, ">>>");
         byte[] toFileContentBytes = Base64.decodeBase64(toFileContentBase64String);
+        File toFile = new File(curPath + toFileName);
+
+        if (toFile.exists())
+        {
+            System.out.println("fromFileName: " + fromFileName);
+            System.out.println("toFileName:   " + toFileName);
+            System.out.println(String.format("File [%s] exists, skip to string2File.", toFileName));
+            System.out.println();
+            return;
+        }
 
         // 写入文件
         OutputStream toFileContentOutputStream = new FileOutputStream(curPath + toFileName);
@@ -55,5 +79,6 @@ public class String2File
 
         System.out.println("fromFileName: " + fromFileName);
         System.out.println("toFileName:   " + toFileName);
+        System.out.println();
     }
 }
